+++ CHANGELOG.md
@@ -10,6 +10,14 @@
 - `procfs` plugin was throwing an exception while parsing
   `/proc/net/dev` to populate `_OP_IPV[4/6]_INTERFACES` keys.
   ([#399](https://github.com/crashappsec/chalk/pull/399))
+- `_IMAGE_DIGEST` is sent when for `docker push` when
+  buildx is not available. Normally chalk needs to validate
+  type of the manifest in the registry (image or list)
+  which is currently done via `buildx imagetools`.
+  When buildx is missing and the operation was `docker push`
+  the pushed image can only be image manifest as only buildx
+  supports list manifests.
+  ([#401](https://github.com/crashappsec/chalk/pull/401))
 
 ## 0.4.10